+++ .github/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### New features
 
-* Implements the pre-Iwazawa and Iwazawa decompositions for symplectic matrices [(#382)](https://github.com/XanaduAI/thewalrus/pull/382).
+* Implements the pre-Iwasawa and Iwasawa decompositions for symplectic matrices [(#382)](https://github.com/XanaduAI/thewalrus/pull/382).
 
 ### Breaking changes